+++ packages/stargate/src/modules/group/aminomessages.ts
@@ -1,0 +1,6 @@
+import { AminoConverters } from "../../aminotypes";
+
+export function createGroupAminoConverters(): AminoConverters {
+  // Missing, see https://github.com/cosmos/cosmjs/issues/1441
+  return {};
+}